+++ dockci/api/fields.py
@@ -5,11 +5,42 @@ marshaling
 
 import re
 
-from functools import wraps
+from functools import reduce, wraps
 
 from flask_restful import fields
 
 from dockci.util import gravatar_url
+
+
+def value_path(obj, path):
+    """
+    Get a value from the given object by dot-separated path
+
+    Examples:
+
+    >>> class TestClass(object):
+    ...     pass
+
+    >>> testobj = TestClass()
+    >>> value_path(testobj, 'testval.teststr')
+    Traceback (most recent call last):
+        ...
+    AttributeError: 'TestClass' object has no attribute 'testval'
+
+    >>> testobj.testval = TestClass()
+    >>> testobj.testval.teststr = 'Test okay'
+    >>> value_path(testobj, 'testval.teststr')
+    'Test okay'
+
+    >>> testobj.testval = None
+    >>> value_path(testobj, 'testval.teststr')
+    """
+    return reduce(
+        lambda acc, attr_name:
+            None if acc is None else getattr(acc, attr_name),
+        path.split('.'),
+        obj,
+    )
 
 
 class RewriteUrl(fields.Url):
@@ -31,13 +62,7 @@ class RewriteUrl(fields.Url):
 
         data = obj.__dict__
         for field_set, field_from in self.rewrites.items():
-            attr_path_data = obj
-            for attr_path in field_from.split('.'):
-                if attr_path_data is None:
-                    return None
-                attr_path_data = getattr(attr_path_data, attr_path)
-
-            data[field_set] = attr_path_data
+            data[field_set] = value_path(obj, field_from)
 
         return super(RewriteUrl, self).output(key, data)
 
@@ -46,11 +71,16 @@ class GravatarUrl(fields.String):
     """
     Automatically turn an email into a Gravatar URL
 
+    >>> from dockci.models.auth import User, UserEmail
     >>> from dockci.models.job import Job
 
     >>> field = GravatarUrl()
     >>> field.output('git_author_email',
     ...              Job(git_author_email='ricky@spruce.sh'))
+    'https://s.gravatar.com/avatar/35866d5d838f7aeb9b51a29eda9878e7'
+
+    >>> field.output('email_obj.email',
+    ...              User(email_obj=UserEmail(email='ricky@spruce.sh')))
     'https://s.gravatar.com/avatar/35866d5d838f7aeb9b51a29eda9878e7'
 
     >>> field = GravatarUrl(attr_name='git_author_email')
@@ -60,16 +90,19 @@ class GravatarUrl(fields.String):
 
     >>> field.output('git_author_email',
     ...              Job(git_author_email=None))
+
+    >>> field = GravatarUrl(attr_name='email_obj.email')
+    >>> field.output('different_name',
+    ...              User(email_obj=UserEmail(email='ricky@spruce.sh')))
+    'https://s.gravatar.com/avatar/35866d5d838f7aeb9b51a29eda9878e7'
     """
     def __init__(self, attr_name=None):
         super(GravatarUrl, self).__init__()
         self.attr_name = attr_name
 
     def output(self, key, obj):
-        if self.attr_name is None:
-            email = getattr(obj, key)
-        else:
-            email = getattr(obj, self.attr_name)
+        path = key if self.attr_name is None else self.attr_name
+        email = value_path(obj, path)
 
         if email is None:
             return None