+++ dockci/api/user.py
@@ -1,12 +1,13 @@
 """ API relating to User model objects """
+from flask import abort
 from flask_restful import fields, inputs, marshal_with, Resource
 from flask_security import current_user, login_required
 
 from .base import BaseDetailResource, BaseRequestParser
 from .fields import GravatarUrl, NonBlankInput, RewriteUrl
 from .util import DT_FORMATTER, new_edit_parsers
-from dockci.models.auth import User
-from dockci.server import API
+from dockci.models.auth import User, UserEmail
+from dockci.server import API, APP, DB
 
 
 BASIC_FIELDS = {
@@ -24,6 +25,7 @@ LIST_FIELDS.update(BASIC_FIELDS)
 DETAIL_FIELDS = {
     'avatar': GravatarUrl(attr_name='email'),
     'confirmed_at': DT_FORMATTER,
+    'emails': fields.List(fields.String()),
 }
 DETAIL_FIELDS.update(BASIC_FIELDS)
 
@@ -48,6 +50,9 @@ USER_EDIT_PARSER.add_argument('active',
                               type=inputs.boolean)
 
 
+SECURITY_STATE = APP.extensions['security']
+
+
 # pylint:disable=no-self-use
 
 class UserList(BaseDetailResource):
@@ -67,18 +72,28 @@ class UserList(BaseDetailResource):
 
 class UserDetail(BaseDetailResource):
     """ API resource that handles getting user details, and updating users """
+    @classmethod
+    def user_or_404(cls, user_id):
+        """ Return a user from the security store, or 404 """
+        user = SECURITY_STATE.datastore.get_user(user_id)
+        if user is None:
+            return abort(404)
+
+        return user
+
     @login_required
     @marshal_with(DETAIL_FIELDS)
     def get(self, user_id):
         """ Get a user's details """
-        return User.query.get_or_404(user_id)
+        return self.user_or_404(user_id)
 
     @login_required
     @marshal_with(DETAIL_FIELDS)
     def post(self, user_id, user=None):
         """ Update a user """
         if user is None:
-            user = User.query.get_or_404(user_id)
+            user = self.user_or_404(user_id)
+
         return self.handle_write(user, USER_EDIT_PARSER)
 
 
@@ -96,6 +111,19 @@ class MeDetail(Resource):
         return UserDetail().post(None, current_user)
 
 
+class MeEmailDetail(Resource):
+    """ Deletion of user email addresses """
+    @login_required
+    def delete(self, email):
+        """ Delete an email from the current user """
+        email = current_user.emails.filter(
+            UserEmail.email.ilike(email),
+        ).first_or_404()
+        DB.session.delete(email)
+        DB.session.commit()
+        return {'message': '%s deleted' % email.email}
+
+
 API.add_resource(UserList,
                  '/users',
                  endpoint='user_list')
@@ -105,3 +133,6 @@ API.add_resource(UserDetail,
 API.add_resource(MeDetail,
                  '/me',
                  endpoint='me_detail')
+API.add_resource(MeEmailDetail,
+                 '/me/<string:email>',
+                 endpoint='me_email_detail')